# examples/apiary/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Apiary API",
    "description": "Manage beehives: count bees, move colonies and harvest honey. An example of swaggerkit.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
    "/hives": {
      "get": {
        "tags": [
          "hives"
        ],
        "summary": "List hives",
        "operationId": "listHives",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Page size",
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Items to skip",
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Only hives in this state",
            "schema": {
              "$ref": "#/components/schemas/HiveStatus"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "description": "Only hives with all of these tags",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "maxItems": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageHive"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "hives"
        ],
        "summary": "Create hive",
        "operationId": "createHive",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewHive"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hive"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "beekeeper": []
          }
        ]
      }
    },
    "/hives/{hiveId}": {
      "delete": {
        "tags": [
          "hives"
        ],
        "summary": "Delete hive",
        "operationId": "deleteHive",
        "parameters": [
          {
            "name": "hiveId",
            "in": "path",
            "description": "Hive ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "examples": [
                1
              ],
              "minimum": 1
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "beekeeper": []
          }
        ]
      },
      "get": {
        "tags": [
          "hives"
        ],
        "summary": "Get hive",
        "operationId": "getHive",
        "parameters": [
          {
            "name": "hiveId",
            "in": "path",
            "description": "Hive ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "examples": [
                1
              ],
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hive"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "hives"
        ],
        "summary": "Update hive",
        "operationId": "updateHive",
        "parameters": [
          {
            "name": "hiveId",
            "in": "path",
            "description": "Hive ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "examples": [
                1
              ],
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HiveUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hive"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "beekeeper": []
          }
        ]
      }
    },
    "/hives/{hiveId}/bees": {
      "post": {
        "tags": [
          "hives"
        ],
        "summary": "Move bees into a hive",
        "operationId": "addBees",
        "parameters": [
          {
            "name": "hiveId",
            "in": "path",
            "description": "Hive ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "examples": [
                1
              ],
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "count": {
                    "type": "integer",
                    "format": "int64",
                    "description": "Number of bees to add",
                    "examples": [
                      5000
                    ],
                    "minimum": 1,
                    "maximum": 50000
                  }
                },
                "required": [
                  "count"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hive"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "beekeeper": []
          }
        ]
      }
    },
    "/hives/{hiveId}/harvests": {
      "post": {
        "tags": [
          "hives",
          "honey"
        ],
        "summary": "Harvest honey",
        "operationId": "harvestHoney",
        "parameters": [
          {
            "name": "hiveId",
            "in": "path",
            "description": "Hive ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "examples": [
                1
              ],
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "kg": {
                    "type": "number",
                    "format": "double",
                    "description": "Honey to take, kg",
                    "examples": [
                      4.5
                    ],
                    "maximum": 100,
                    "exclusiveMinimum": 0
                  }
                },
                "required": [
                  "kg"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Harvest"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "beekeeper": []
          }
        ]
      }
    },
    "/hives/{hiveId}/label": {
      "get": {
        "tags": [
          "hives"
        ],
        "summary": "Print a hive label",
        "operationId": "hiveLabel",
        "parameters": [
          {
            "name": "hiveId",
            "in": "path",
            "description": "Hive ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "examples": [
                1
              ],
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/hives/{hiveId}/photo": {
      "get": {
        "tags": [
          "hives"
        ],
        "summary": "Get the photo of a hive",
        "operationId": "getPhoto",
        "parameters": [
          {
            "name": "hiveId",
            "in": "path",
            "description": "Hive ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "examples": [
                1
              ],
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "image/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "hives"
        ],
        "summary": "Upload a photo of a hive",
        "operationId": "uploadPhoto",
        "parameters": [
          {
            "name": "hiveId",
            "in": "path",
            "description": "Hive ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "examples": [
                1
              ],
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "caption": {
                    "type": "string",
                    "description": "Caption",
                    "maxLength": 100
                  },
                  "photo": {
                    "type": "string",
                    "format": "binary",
                    "description": "JPEG or PNG image, up to 2 MiB"
                  }
                },
                "required": [
                  "photo"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoInfo"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "beekeeper": []
          }
        ]
      }
    },
    "/stats": {
      "get": {
        "tags": [
          "apiary"
        ],
        "summary": "Count hives, bees and honey",
        "operationId": "getStats",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Stats"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Harvest": {
        "type": "object",
        "properties": {
          "beekeeper": {
            "type": "string",
            "description": "Who harvested",
            "examples": [
              "beekeeper"
            ]
          },
          "harvestedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the honey was taken"
          },
          "hiveId": {
            "type": "integer",
            "format": "int64",
            "description": "Hive the honey was taken from",
            "examples": [
              1
            ]
          },
          "kg": {
            "type": "number",
            "format": "double",
            "description": "Honey taken, kg",
            "examples": [
              4.5
            ]
          },
          "remainingKg": {
            "type": "number",
            "format": "double",
            "description": "Honey left in the hive, kg",
            "examples": [
              8
            ]
          }
        },
        "required": [
          "hiveId",
          "kg",
          "remainingKg",
          "harvestedAt",
          "beekeeper"
        ]
      },
      "Hive": {
        "type": "object",
        "properties": {
          "bees": {
            "type": "integer",
            "format": "int64",
            "description": "Number of bees living in the hive",
            "examples": [
              42000
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the hive was set up"
          },
          "honeyKg": {
            "type": "number",
            "format": "double",
            "description": "Honey stored in the hive, kg",
            "examples": [
              12.5
            ]
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Hive ID",
            "examples": [
              1
            ]
          },
          "location": {
            "$ref": "#/components/schemas/Location",
            "description": "Where the hive stands"
          },
          "name": {
            "type": "string",
            "description": "Name painted on the hive",
            "examples": [
              "Linden"
            ]
          },
          "queen": {
            "$ref": "#/components/schemas/Queen",
            "description": "Queen of the colony, if there is one"
          },
          "status": {
            "$ref": "#/components/schemas/HiveStatus",
            "description": "Colony state"
          },
          "tags": {
            "type": "array",
            "description": "Free-form labels",
            "examples": [
              [
                "meadow",
                "calm"
              ]
            ],
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "bees",
          "honeyKg",
          "location",
          "tags",
          "createdAt"
        ]
      },
      "HiveStatus": {
        "type": "string",
        "enum": [
          "active",
          "swarming",
          "dormant",
          "empty"
        ]
      },
      "HiveUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "New name",
            "minLength": 1,
            "maxLength": 64
          },
          "queen": {
            "$ref": "#/components/schemas/Queen",
            "description": "New queen"
          },
          "status": {
            "$ref": "#/components/schemas/HiveStatus",
            "description": "New colony state"
          }
        }
      },
...
      "NewHive": {
        "type": "object",
        "properties": {
          "location": {
            "$ref": "#/components/schemas/Location",
            "description": "Where the hive stands"
          },
          "name": {
            "type": "string",
            "description": "Name painted on the hive",
            "examples": [
              "Linden"
            ],
            "minLength": 1,
            "maxLength": 64
          },
          "status": {
            "$ref": "#/components/schemas/HiveStatus",
            "description": "Initial colony state",
            "default": "empty"
          },
          "tags": {
            "type": "array",
            "description": "Free-form labels",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 32
            },
            "maxItems": 10,
            "uniqueItems": true
          }
        },
        "required": [
          "name",
          "location"
        ]
      },
      "PageHive": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "description": "Items on this page",
            "items": {
              "$ref": "#/components/schemas/Hive"
            }
          },
          "limit": {
            "type": "integer",
            "format": "int64",
            "description": "Page size",
            "examples": [
              20
            ]
          },
          "offset": {
            "type": "integer",
            "format": "int64",
            "description": "Items skipped",
            "examples": [
              0
            ]
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "Number of items in all pages",
            "examples": [
              3
            ]
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
      "PhotoInfo": {
        "type": "object",
        "properties": {
          "caption": {
            "type": "string",
            "description": "Caption",
            "examples": [
              "Spring inspection"
            ]
          },
          "contentType": {
            "type": "string",
            "description": "image/jpeg or image/png",
            "examples": [
              "image/jpeg"
            ]
          },
          "hiveId": {
            "type": "integer",
            "format": "int64",
            "description": "Hive ID",
            "examples": [
              1
            ]
          },
          "name": {
            "type": "string",
            "description": "Original file name",
            "examples": [
              "linden.jpg"
            ]
          },
          "size": {
            "type": "integer",
            "format": "int64",
            "description": "Size in bytes",
            "examples": [
              48213
            ]
          }
        },
        "required": [
          "hiveId",
          "name",
          "contentType",
          "size"
        ]
      },
      "Problem": {
        "type": "object",
        "description": "Problem details, RFC 9457.",
        "properties": {
          "detail": {
            "type": "string",
            "examples": [
              "request validation failed"
            ]
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "location": {
                  "type": "string",
                  "examples": [
                    "body.name"
                  ]
                },
                "message": {
                  "type": "string",
                  "examples": [
                    "is required"
                  ]
                }
              },
              "required": [
                "location",
                "message"
              ]
            }
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "examples": [
              422
            ]
          },
          "title": {
            "type": "string",
            "examples": [
              "Unprocessable Entity"
            ]
          },
          "type": {
            "type": "string",
            "format": "uri-reference",
            "examples": [
              "about:blank"
            ]
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ]
      },
...
      "Stats": {
        "type": "object",
        "properties": {
          "bees": {
            "type": "integer",
            "format": "int64",
            "description": "Bees in all hives",
            "examples": [
              97000
            ]
          },
          "hives": {
            "type": "integer",
            "format": "int64",
            "description": "Number of hives",
            "examples": [
              3
            ]
          },
          "honeyKg": {
            "type": "number",
            "format": "double",
            "description": "Honey in all hives, kg",
            "examples": [
              31.5
            ]
          }
        },
        "required": [
          "hives",
          "bees",
          "honeyKg"
        ]
      }
    },
    "securitySchemes": {
      "beekeeper": {
        "type": "http",
        "description": "Beekeeper token. The demo token is `beekeeper`.",
        "scheme": "bearer"
      }
    }
  }
}